Plausibly working version
See <file:abcl/contrib/abcl-stepper/steppenwolf.org> for a recipe to
get this working from the REPL.
