+++ contrib/abcl-stepper/abcl-stepper.asd
@@ -3,8 +3,11 @@
   :author "Alejandro Zamora Fonseca"
   :description "An operational stepper for ABCL"
   :long-description "<urn:abcl.org/release/1.9.1/contrib/abcl-stepper#>"
-  :version "0.0.1"
-  :components ((:module base
-                        :pathname ""
-                        :components ((:file "abcl-stepper")
-                                     (:static-file "README.markdown")))))
+  :version "0.0.2"
+  :depends-on (steppenwolf)
+  :components  ((:module package :pathname "./"
+                 :components ((:file "package")))
+                ((:module base :pathname "./"
+                  :depends-on (package)
+                  :components ((:file "abcl-stepper")
+                               (:static-file "README.markdown"))))))
