+++ src/org/armedbear/lisp/Lisp.java
@@ -39,90 +39,97 @@ import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Reader;
 import java.io.StringReader;
+import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.math.BigInteger;
 import java.net.URL;
 import java.nio.charset.Charset;
 import java.util.Hashtable;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 public final class Lisp
 {
+  // TODO remove me!
+  public static Method stepperHook
+    = null;
+  
   public static final boolean debug = true;
 
   public static boolean cold = true;
 
   public static boolean initialized;
 
-  // stepping related attributes
-  public static boolean stepping = false;
-  public static boolean delimitedStepping = false;
-  public static boolean firstStepping = false;
-  public static Binding stepperBlock = null;
+  // // stepping related attributes
+  // public static boolean stepping = false;
+  // public static boolean delimitedStepping = false;
+  // public static boolean firstStepping = false;
+  // public static Binding stepperBlock = null;
 
-  public static boolean stepInSymbolP (LispObject function) {
-    Package stepper;
-    Symbol symbol;
-    LispThread currentThread = LispThread.currentThread();
-    LispObject stopAtSymbolFunction;
-    LispObject[] args = new LispObject[1];
-    LispObject result;
+  // public static boolean stepInSymbolP (LispObject function) {
+  //   Package stepper;
+  //   Symbol symbol;
+  //   LispThread currentThread = LispThread.currentThread();
+  //   LispObject stopAtSymbolFunction;
+  //   LispObject[] args = new LispObject[1];
+  //   LispObject result;
 
-    if (stepping && firstStepping) {
-      firstStepping = false;
-      stepperBlock = null;
-      return false;
-    }
+  //   if (stepping && firstStepping) {
+  //     firstStepping = false;
+  //     stepperBlock = null;
+  //     return false;
+  //   }
 
-    if (!stepping && !delimitedStepping) {
-      return false;
-    }
+  //   if (!stepping && !delimitedStepping) {
+  //     return false;
+  //   }
 
-    if (stepping && !delimitedStepping) {
-      return true;
-    }
+  //   if (stepping && !delimitedStepping) {
+  //     return true;
+  //   }
 
-    if (stepping && delimitedStepping) {
-      // we analyze if the symbol of the function is one of the
-      // symbols exported in the packages specified in sys::*stop-packages*
-      // or one of the symbols in sys::*stop-symbols*
-      // in that case we trigger the stepper here
-      setSteppingOff();
-      stepper = Packages.findPackageGlobally("ABCL-STEPPER");
-      symbol = stepper.findAccessibleSymbol("STOP-AT-SYMBOL-P");
-      stopAtSymbolFunction = coerceToFunction(symbol);
-      args[0] = ((Operator)function).getLambdaName();
-      result = funcall(stopAtSymbolFunction, args, currentThread);
-      setSteppingOnAfterInit();
-      if (result != NIL) {
-        setDelimitedSteppingOff();
-        return true;
-      }
-      return false;
-    }
-    return false;
-  }
+  //   if (stepping && delimitedStepping) {
+  //     // we analyze if the symbol of the function is one of the
+  //     // symbols exported in the packages specified in sys::*stop-packages*
+  //     // or one of the symbols in sys::*stop-symbols*
+  //     // in that case we trigger the stepper here
+  //     setSteppingOff();
+  //     stepper = Packages.findPackageGlobally("ABCL-STEPPER");
+  //     symbol = stepper.findAccessibleSymbol("STOP-AT-SYMBOL-P");
+  //     stopAtSymbolFunction = coerceToFunction(symbol);
+  //     args[0] = ((Operator)function).getLambdaName();
+  //     result = funcall(stopAtSymbolFunction, args, currentThread);
+  //     setSteppingOnAfterInit();
+  //     if (result != NIL) {
+  //       setDelimitedSteppingOff();
+  //       return true;
+  //     }
+  //     return false;
+  //   }
+  //   return false;
+  // }
 
-  public static void setDelimitedSteppingOn () {
-    delimitedStepping = true;
-  }
+  // public static void setDelimitedSteppingOn () {
+  //   delimitedStepping = true;
+  // }
 
-  public static void setDelimitedSteppingOff () {
-    delimitedStepping = false;
-  }
+  // public static void setDelimitedSteppingOff () {
+  //   delimitedStepping = false;
+  // }
 
-  public static void setSteppingOn () {
-    stepping = true;
-    firstStepping = true;
-  }
+  // public static void setSteppingOn () {
+  //   stepping = true;
+  //   firstStepping = true;
+  // }
 
-  public static void setSteppingOnAfterInit () {
-    stepping = true;
-  }
+  // public static void setSteppingOnAfterInit () {
+  //   stepping = true;
+  // }
 
-  public static void setSteppingOff () {
-    stepping = false;
-  }
+  // public static void setSteppingOff () {
+  //   stepping = false;
+  // }
 
   // Packages.
   public static final Package PACKAGE_CL =
@@ -586,145 +593,145 @@ public static synchronized final void handleInterrupt()
   }
 
 
-  public static final void printForStepping(LispObject printFunction, SimpleString str, LispThread currentThread, boolean newline) {
-    LispObject [] argsPrint = new LispObject[2];
-    argsPrint[0] = str;
-    argsPrint[1] = newline? T : NIL;
-    funcall(printFunction, argsPrint, currentThread);
-  }
+  // public static final void printForStepping(LispObject printFunction, SimpleString str, LispThread currentThread, boolean newline) {
+  //   LispObject [] argsPrint = new LispObject[2];
+  //   argsPrint[0] = str;
+  //   argsPrint[1] = newline? T : NIL;
+  //   funcall(printFunction, argsPrint, currentThread);
+  // }
 
-  public static synchronized final void handleStepping (LispObject function, LispObject args,
-                                                        Environment env) {
+  // public static synchronized final void handleStepping (LispObject function, LispObject args,
+  //                                                       Environment env) {
 
-    LispThread currentThread = LispThread.currentThread();
-    Package stepper = Packages.findPackageGlobally("ABCL-STEPPER");
-    Package system = Packages.findPackageGlobally("SYSTEM");
-    Symbol symbolPrintStepperStr = stepper.findAccessibleSymbol("PRINT-STEPPER-STR");
-    LispObject functionPrintStepperStr = coerceToFunction(symbolPrintStepperStr);
-    if (stepperBlock == null) {
-      stepperBlock = env.getOuterMostBlock();
-    }
-    printForStepping(functionPrintStepperStr, new SimpleString("We are in the stepper mode"), currentThread, true);
-    printForStepping(functionPrintStepperStr, new SimpleString("Evaluating :"), currentThread, true);
-    LispObject closureName = ((Operator)function).getLambdaName();
-    if (closureName != null ) {
-      printForStepping(functionPrintStepperStr, new SimpleString(closureName.printObject()), currentThread, true);
-    }
-    else {
-      printForStepping(functionPrintStepperStr, new SimpleString(((Operator)function).printObject()), currentThread, true);
-    }
-    printForStepping(functionPrintStepperStr, new SimpleString("With args: "), currentThread, true);
-    if (args == NIL) {
-      printForStepping(functionPrintStepperStr, new SimpleString("NIL"), currentThread, true);
-    }
-    else {
-      printForStepping(functionPrintStepperStr, new SimpleString(args.printObject()), currentThread, true);
-    }
-    boolean leavePrompt = false;
-    boolean unexpectedInputUser = false;
-    while (!leavePrompt) {
-      if (!unexpectedInputUser) {
-        printForStepping(functionPrintStepperStr, new SimpleString("Type '?' for a list of options"), currentThread, true);
-      }
-      unexpectedInputUser = false;
-      LispObject charInputUser = funcall(coerceToFunction(Symbol.READ_CHAR), new LispObject[]{}, currentThread);
-      funcall(coerceToFunction(Symbol.CLEAR_INPUT), new LispObject[]{}, currentThread);
-      char inputUser = ((LispCharacter)charInputUser).value;
-      String variableName;
-      String packageName;
-      Package pkg;
-      switch (inputUser) {
-      case '?':
-        setSteppingOff();
-        Symbol symbolPrintStepperHelp = stepper.findAccessibleSymbol("PRINT-STEPPER-HELP");
-        LispObject functionPrintStepperHelp = coerceToFunction(symbolPrintStepperHelp);
-        funcall(functionPrintStepperHelp, new LispObject[]{}, currentThread);
-        setSteppingOnAfterInit();
-        break;
-      case 'l':
-        setSteppingOff();
-        Symbol symbolVariables = system.findAccessibleSymbol("ENVIRONMENT-ALL-VARIABLES");
-        Symbol symbolFunctions = system.findAccessibleSymbol("ENVIRONMENT-ALL-FUNCTIONS");
-        LispObject environmentAllVariables = coerceToFunction(symbolVariables);
-        LispObject environmentAllFunctions = coerceToFunction(symbolFunctions);
-        LispObject[] argsEnv = new LispObject[1];
-        argsEnv[0] = env;
-        LispObject resultVars = funcall(environmentAllVariables, argsEnv, currentThread);
-        Symbol symbolPprintListLocals = stepper.findAccessibleSymbol("PPRINT-LIST-LOCALS");
-        LispObject functionPprintListLocals = coerceToFunction(symbolPprintListLocals);
-        printForStepping(functionPrintStepperStr, new SimpleString("Showing the values of variable bindings."), currentThread, true);
-        printForStepping(functionPrintStepperStr, new SimpleString("From inner to outer scopes:"), currentThread, true);
-        argsEnv[0] = resultVars;
-        funcall(functionPprintListLocals, argsEnv, currentThread);
-        argsEnv[0] = env;
-        LispObject resultFuncs = funcall(environmentAllFunctions, argsEnv, currentThread);
-        printForStepping(functionPrintStepperStr, new SimpleString("Showing the values of functions bindings."), currentThread, true);
-        printForStepping(functionPrintStepperStr, new SimpleString("From inner to outer scopes:"), currentThread, true);
-        argsEnv[0] = resultFuncs;
-        funcall(functionPprintListLocals, argsEnv, currentThread);
-        setSteppingOnAfterInit();
-        break;
-      case 'c':
-        setSteppingOff();
-        leavePrompt = true;
-        break;
-      case 'n':
-        setDelimitedSteppingOn();
-        leavePrompt = true;
-        break;
-      case 's':
-        leavePrompt = true;
-        break;
-      case 'q':
-        setSteppingOff();
-        setDelimitedSteppingOff();
-        throw new Return(stepperBlock.symbol, stepperBlock.value, NIL);
-      case 'i':
-        printForStepping(functionPrintStepperStr, new SimpleString("Type the name of the symbol: "), currentThread, false);
-        LispObject variableNameLisp = funcall(coerceToFunction(Symbol.READ_LINE), new LispObject[]{}, currentThread);
-        variableName = variableNameLisp.toString().toUpperCase();
-        printForStepping(functionPrintStepperStr, new SimpleString("Type the name of the package ('-' for current package): "), currentThread, false);
-        LispObject packageNameLisp = funcall(coerceToFunction(Symbol.READ_LINE), new LispObject[]{}, currentThread);
-        packageName = packageNameLisp.toString().toUpperCase();
-        if (packageName.equals("-")) {
-          pkg = getCurrentPackage();
-        }
-        else {
-          pkg = Packages.findPackageGlobally(packageName);
-        }
-        if (pkg == null) {
-          printForStepping(functionPrintStepperStr, new SimpleString(String.format("Couldn't find the package: %s", packageName)), currentThread, true);
-        }
-        else {
-          Symbol symbol = pkg.findAccessibleSymbol(variableName);
-          if (symbol == null) {
-            String message = String.format("Couldn't find the symbol %s in the package %s", variableName, pkg.getName());
-            printForStepping(functionPrintStepperStr, new SimpleString(message), currentThread, true);
-          }
-          else {
-            LispObject variableValue = env.lookup(symbol);
-            if (variableValue == null) {
-              LispObject symbolValue = ((Symbol)symbol).symbolValueNoThrow();
-              if (symbolValue != null) {
-                printForStepping(functionPrintStepperStr, new SimpleString(symbolValue.printObject()), currentThread, true);
-              }
-              else {
-                printForStepping(functionPrintStepperStr,
-                                 new SimpleString(String.format("Couldn't find the value for: %s", variableName)),
-                                 currentThread, true);
-              }
-            }
-            else {
-              printForStepping(functionPrintStepperStr, new SimpleString(variableValue.printObject()), currentThread, true);
-            }
-          }
-        }
-        break;
-      default:
-        unexpectedInputUser = true;
-      }
-    }
-  }
+  //   LispThread currentThread = LispThread.currentThread();
+  //   Package stepper = Packages.findPackageGlobally("ABCL-STEPPER");
+  //   Package system = Packages.findPackageGlobally("SYSTEM");
+  //   Symbol symbolPrintStepperStr = stepper.findAccessibleSymbol("PRINT-STEPPER-STR");
+  //   LispObject functionPrintStepperStr = coerceToFunction(symbolPrintStepperStr);
+  //   if (stepperBlock == null) {
+  //     stepperBlock = env.getOuterMostBlock();
+  //   }
+  //   printForStepping(functionPrintStepperStr, new SimpleString("We are in the stepper mode"), currentThread, true);
+  //   printForStepping(functionPrintStepperStr, new SimpleString("Evaluating :"), currentThread, true);
+  //   LispObject closureName = ((Operator)function).getLambdaName();
+  //   if (closureName != null ) {
+  //     printForStepping(functionPrintStepperStr, new SimpleString(closureName.printObject()), currentThread, true);
+  //   }
+  //   else {
+  //     printForStepping(functionPrintStepperStr, new SimpleString(((Operator)function).printObject()), currentThread, true);
+  //   }
+  //   printForStepping(functionPrintStepperStr, new SimpleString("With args: "), currentThread, true);
+  //   if (args == NIL) {
+  //     printForStepping(functionPrintStepperStr, new SimpleString("NIL"), currentThread, true);
+  //   }
+  //   else {
+  //     printForStepping(functionPrintStepperStr, new SimpleString(args.printObject()), currentThread, true);
+  //   }
+  //   boolean leavePrompt = false;
+  //   boolean unexpectedInputUser = false;
+  //   while (!leavePrompt) {
+  //     if (!unexpectedInputUser) {
+  //       printForStepping(functionPrintStepperStr, new SimpleString("Type '?' for a list of options"), currentThread, true);
+  //     }
+  //     unexpectedInputUser = false;
+  //     LispObject charInputUser = funcall(coerceToFunction(Symbol.READ_CHAR), new LispObject[]{}, currentThread);
+  //     funcall(coerceToFunction(Symbol.CLEAR_INPUT), new LispObject[]{}, currentThread);
+  //     char inputUser = ((LispCharacter)charInputUser).value;
+  //     String variableName;
+  //     String packageName;
+  //     Package pkg;
+  //     switch (inputUser) {
+  //     case '?':
+  //       setSteppingOff();
+  //       Symbol symbolPrintStepperHelp = stepper.findAccessibleSymbol("PRINT-STEPPER-HELP");
+  //       LispObject functionPrintStepperHelp = coerceToFunction(symbolPrintStepperHelp);
+  //       funcall(functionPrintStepperHelp, new LispObject[]{}, currentThread);
+  //       setSteppingOnAfterInit();
+  //       break;
+  //     case 'l':
+  //       setSteppingOff();
+  //       Symbol symbolVariables = system.findAccessibleSymbol("ENVIRONMENT-ALL-VARIABLES");
+  //       Symbol symbolFunctions = system.findAccessibleSymbol("ENVIRONMENT-ALL-FUNCTIONS");
+  //       LispObject environmentAllVariables = coerceToFunction(symbolVariables);
+  //       LispObject environmentAllFunctions = coerceToFunction(symbolFunctions);
+  //       LispObject[] argsEnv = new LispObject[1];
+  //       argsEnv[0] = env;
+  //       LispObject resultVars = funcall(environmentAllVariables, argsEnv, currentThread);
+  //       Symbol symbolPprintListLocals = stepper.findAccessibleSymbol("PPRINT-LIST-LOCALS");
+  //       LispObject functionPprintListLocals = coerceToFunction(symbolPprintListLocals);
+  //       printForStepping(functionPrintStepperStr, new SimpleString("Showing the values of variable bindings."), currentThread, true);
+  //       printForStepping(functionPrintStepperStr, new SimpleString("From inner to outer scopes:"), currentThread, true);
+  //       argsEnv[0] = resultVars;
+  //       funcall(functionPprintListLocals, argsEnv, currentThread);
+  //       argsEnv[0] = env;
+  //       LispObject resultFuncs = funcall(environmentAllFunctions, argsEnv, currentThread);
+  //       printForStepping(functionPrintStepperStr, new SimpleString("Showing the values of functions bindings."), currentThread, true);
+  //       printForStepping(functionPrintStepperStr, new SimpleString("From inner to outer scopes:"), currentThread, true);
+  //       argsEnv[0] = resultFuncs;
+  //       funcall(functionPprintListLocals, argsEnv, currentThread);
+  //       setSteppingOnAfterInit();
+  //       break;
+  //     case 'c':
+  //       setSteppingOff();
+  //       leavePrompt = true;
+  //       break;
+  //     case 'n':
+  //       setDelimitedSteppingOn();
+  //       leavePrompt = true;
+  //       break;
+  //     case 's':
+  //       leavePrompt = true;
+  //       break;
+  //     case 'q':
+  //       setSteppingOff();
+  //       setDelimitedSteppingOff();
+  //       throw new Return(stepperBlock.symbol, stepperBlock.value, NIL);
+  //     case 'i':
+  //       printForStepping(functionPrintStepperStr, new SimpleString("Type the name of the symbol: "), currentThread, false);
+  //       LispObject variableNameLisp = funcall(coerceToFunction(Symbol.READ_LINE), new LispObject[]{}, currentThread);
+  //       variableName = variableNameLisp.toString().toUpperCase();
+  //       printForStepping(functionPrintStepperStr, new SimpleString("Type the name of the package ('-' for current package): "), currentThread, false);
+  //       LispObject packageNameLisp = funcall(coerceToFunction(Symbol.READ_LINE), new LispObject[]{}, currentThread);
+  //       packageName = packageNameLisp.toString().toUpperCase();
+  //       if (packageName.equals("-")) {
+  //         pkg = getCurrentPackage();
+  //       }
+  //       else {
+  //         pkg = Packages.findPackageGlobally(packageName);
+  //       }
+  //       if (pkg == null) {
+  //         printForStepping(functionPrintStepperStr, new SimpleString(String.format("Couldn't find the package: %s", packageName)), currentThread, true);
+  //       }
+  //       else {
+  //         Symbol symbol = pkg.findAccessibleSymbol(variableName);
+  //         if (symbol == null) {
+  //           String message = String.format("Couldn't find the symbol %s in the package %s", variableName, pkg.getName());
+  //           printForStepping(functionPrintStepperStr, new SimpleString(message), currentThread, true);
+  //         }
+  //         else {
+  //           LispObject variableValue = env.lookup(symbol);
+  //           if (variableValue == null) {
+  //             LispObject symbolValue = ((Symbol)symbol).symbolValueNoThrow();
+  //             if (symbolValue != null) {
+  //               printForStepping(functionPrintStepperStr, new SimpleString(symbolValue.printObject()), currentThread, true);
+  //             }
+  //             else {
+  //               printForStepping(functionPrintStepperStr,
+  //                                new SimpleString(String.format("Couldn't find the value for: %s", variableName)),
+  //                                currentThread, true);
+  //             }
+  //           }
+  //           else {
+  //             printForStepping(functionPrintStepperStr, new SimpleString(variableValue.printObject()), currentThread, true);
+  //           }
+  //         }
+  //       }
+  //       break;
+  //     default:
+  //       unexpectedInputUser = true;
+  //     }
+  //   }
+  // }
 
   public static final LispObject eval(final LispObject obj,
                                       final Environment env,
@@ -776,10 +783,20 @@ public static synchronized final void handleInterrupt()
                 if (profiling)
                   if (!sampling)
                     fun.incrementCallCount();
-                // Don't eval args!
-                if (stepInSymbolP(fun)) {
-                  handleStepping(fun, (obj != NIL) ? ((Cons)obj).cdr : obj, env) ;
+                // TODO remove me!
+                if (stepperHook != null) {
+                  try {
+                    stepperHook.invoke(fun,
+                            (obj != NIL) ? ((Cons)obj).cdr : obj,
+                            env);
+                  } catch (IllegalAccessException ex) {
+                    Logger.getLogger(Lisp.class.getName()).log(Level.SEVERE, null, ex);
+                  } catch (InvocationTargetException ex) {
+                    Logger.getLogger(Lisp.class.getName()).log(Level.SEVERE, null, ex);
+                  }
                 }
+                
+                // Don't eval args!
                 return fun.execute(((Cons)obj).cdr, env);
               }
             if (fun instanceof MacroObject)
@@ -827,8 +844,17 @@ public static synchronized final void handleInterrupt()
                                           LispThread thread)
 
   {
-    if (stepInSymbolP(function)) {
-      handleStepping(function, args != NIL ? ((Cons)args) : args, env);
+    // TODO remove me
+    if (stepperHook != null) {
+      try {
+        stepperHook.invoke(function,
+                           args != NIL ? ((Cons)args) : args,
+                           env);
+      } catch (IllegalAccessException ex) {
+        Logger.getLogger(Lisp.class.getName()).log(Level.SEVERE, null, ex);
+      } catch (InvocationTargetException ex) {
+        Logger.getLogger(Lisp.class.getName()).log(Level.SEVERE, null, ex);
+      }
     }
     if (args == NIL) {
       return thread.execute(function);
@@ -2492,44 +2518,44 @@ public static synchronized final void handleInterrupt()
       }
     };
 
-  // ### %set-stepper-on
-  public static final Primitive SET_STEPPER_ON =
-    new Primitive("%set-stepper-on", PACKAGE_SYS, true)
-    {
-      @Override
-      public LispObject execute()
+  // // ### %set-stepper-on
+  // public static final Primitive SET_STEPPER_ON =
+  //   new Primitive("%set-stepper-on", PACKAGE_SYS, true)
+  //   {
+  //     @Override
+  //     public LispObject execute()
 
-      {
-        setSteppingOn();
-        return NIL;
-      }
-    };
+  //     {
+  //       setSteppingOn();
+  //       return NIL;
+  //     }
+  //   };
 
-  // ### %set-stepper-off
-  public static final Primitive SET_STEPPER_OFF =
-    new Primitive("%set-stepper-off", PACKAGE_SYS, true)
-    {
-      @Override
-      public LispObject execute()
+  // // ### %set-stepper-off
+  // public static final Primitive SET_STEPPER_OFF =
+  //   new Primitive("%set-stepper-off", PACKAGE_SYS, true)
+  //   {
+  //     @Override
+  //     public LispObject execute()
 
-      {
-        setSteppingOff();
-        return NIL;
-      }
-    };
+  //     {
+  //       setSteppingOff();
+  //       return NIL;
+  //     }
+  //   };
 
-  // ### %set-delimited-stepping-off
-  public static final Primitive SET_DELIMITED_STEPPING_OFF =
-    new Primitive("%set-delimited-stepping-off", PACKAGE_SYS, true)
-    {
-      @Override
-      public LispObject execute()
+  // // ### %set-delimited-stepping-off
+  // public static final Primitive SET_DELIMITED_STEPPING_OFF =
+  //   new Primitive("%set-delimited-stepping-off", PACKAGE_SYS, true)
+  //   {
+  //     @Override
+  //     public LispObject execute()
 
-      {
-        setDelimitedSteppingOff();
-        return NIL;
-      }
-    };
+  //     {
+  //       setDelimitedSteppingOff();
+  //       return NIL;
+  //     }
+  //   };
 
   public static final Symbol internSpecial(String name, Package pkg,
                                            LispObject value)
